Task 6 was to create different files for all the functions. I've created a separate cpp-file for each task and one header-file for all. In the main cpp-file only task 6 is presented with a call of all the tasks solutions by pointers on their functions.
