+++ FunctionsAndNamespaces/AllTheFunctions.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include <iostream>
+
+namespace work_with_functions
+{
+	using std::cout;
+
+	void printArray(const bool*, const size_t);
+	void task1();
+
+	void printArray(const int*, const size_t);
+	void task2();
+
+	void task3();
+
+	void task4();
+
+	void task5();
+}
